+++ src/gui/element/wallet/operation/tx_done.rs
@@ -69,7 +69,7 @@ pub fn data_container<'a>(
     let header_container = Container::new(header_row).padding(iced::Padding::from([
         0,               // top
         0,               // right
-        DEFAULT_PADDING, // bottom
+        DEFAULT_PADDING as u16, // bottom
         0,               // left
     ]));
